refactor(runtime-admin): rename Force endpoint to ForceDrain
"Force" alone is meaningless out of context — force what? — and it
sits oddly next to the verb-named siblings Pause / Resume / Status.
The matching admin service method is already IWorkflowRuntimeAdminService.
ForceDrainAsync, so ForceDrain is the natural pair.

Renamed:
- src/modules/Elsa.Workflows.Api/Endpoints/RuntimeAdmin/Force/        → ForceDrain/
- namespace ...Endpoints.RuntimeAdmin.Force                            → ...ForceDrain
- class ForceEndpoint                                                  → ForceDrainEndpoint
- class ForceRequest                                                   → ForceDrainRequest
- class ForceResponse                                                  → ForceDrainResponse
- route  POST /admin/workflow-runtime/force                            → /force-drain

Zero external references — no tests, docs, or OpenAPI clients used the
old symbols or the old route literal, so this is a contained pre-ship
rename. Directory move went through `git mv` so commit history follows
the file.

17/17 graceful-shutdown integration tests still pass.
